app/requirements.txt
databricks-sdk>=0.69.0
flask>=3.0.0
gunicorn>=21.2.0
